work-anniversary-webhook/requirements.txt
gspread
httplib2
google-api-python-client
google-auth
google-cloud-secret-manager